Begin search optimization.
This builds an actual lunr.js search index rather than a JSON array of
page objects which was used to build the index client side.

Note that this is an incomplete solution. This only creates the index.
It is not usable at this point. Just trying to see how well PyExecJS works.