Added nav events.
I don't care much for this. The plugin calls feel like they are in wierd
places and I don't like needing to pass around the PluginCollection
so much. However, outside of a complete refactor of the nav code (seems
unlikely), there is no other way to get access to the plugin events when
they are needed. This feels like an obvious add-on rather than being
built from the ground up with plugin events in mind.

This has given me more pause than anything else regarding the design of
the API. It has been suprisingly simple up to this point.